Fixed gcc warnings.

+++ toxcore/TCP_client.c
@@ -322,7 +322,7 @@ static int write_packet_TCP_secure_connection(TCP_Client_Connection *con, const 
 
         increment_nonce(con->sent_nonce);
 
-        if (len == sizeof(packet)) {
+        if ((unsigned int)len == sizeof(packet)) {
             return 1;
         }
